chatbot_backend/AI/Vikas/requirements.txt
langchain-openai
openai
requests
Django
numpy
pandas
python-dotenv
pytest
python-dotenv
rich
chromadb
langchain_community
pydub
speechrecognition
pyttsx3
pyaudio
fastapi
uvicorn
websockets
fuzzywuzzy
python-Levenshtein
